Fixed|Map Renderer: Dually rendered geometry with rend-dev-sky
When rend-dev-sky was enabled and given a particular map construct
where geometry that would usually be drawable lies inside geometry
that would usually be sky-masked -- both geometries would be drawn.

There is now only one remaining visual glitch with the sky masking
that I am aware of, which is only observable when rend-dev-sky is
enabled. This should also be fixable...