Fixed|Resources: Failed offsetting sky fix planes for two sided wall sections
After the map is loaded the sky fix plane heights are initialized by
scanning the map for certain geometry constructs. However, due to an
initialization order bug in the map setup this logic was ran before
the smoothed surface material origins. Consequently the sky fix was
not being offset when it should be.

Revised Surface so that during map setup any changes to the material
origin are immediately applied to the smoothed coordinates, also.

Updated the sky fix offsetting algorithm to use sector references
from the half-edges rather than map line sides.