Sector: Enumerate referencing map line sides instead of lines
Maintaining a list of map lines which reference the sector is useful.
However, more useful is a list of the referencing map line sides, as
this then simplifies the logic for any users of this list (there is
no need to then deduce which side(s) of the line has the reference).