Fixed: Use the appropriate map line method when querying sectors/sections
The DENG map format separates the dual meaning of a two-sided line
in the id Tech 1 format. A line with a sector on both sides may not
necessarily have surface sections on both sides.

Also, ensuring to use the semantically correct accessor method(s)
when querying the sector(s) and/or sections of a line often removes
the need to implement special case handling for various map hacks
(e.g., one-way windows).