BSP Builder: Produce a full complement of half-edges
The BSP builder now produces two half-edges for every line segment
in the map, regardless of whether a map line has a sector on both
sides. The "additional" to vertex referenced by HEdge has now been
removed.

As of this commit the half-edge data structure is essentially now
implemented and in active use. The next step is to abstract each
level of functionality into a class hierarchy to cleanly separate
the Euler topology/DCEL and map/game concerns.