Fixed: Map geometry generation near degenerate BspLeafs
In cases where degenerate BspLeafs are the result of the map author
defining superfluous back SideDefs in id Tech 1 format maps (such as
in the first map of Aliens TC), the resultant geometry would leave
holes in the map. If a wall construct has a degenerate BspLeaf on
either side of the map line then it should always be treated as if
no relative back sector is defined.

A couple of holes in the first map of Aliens TC, however, as these
are the result of other more serious issues (e.g., where map sectors
partially overlap).