BSP Builder: Discard degenerate geometries produced by the partitioner
BspLeaf now owns a Polygon which it guarantees as both simple and
convex. This separates the concerns of a leaf node in a Binary Tree
with the associated convex geometry.

This in turn allowed revising how degenerate polygon geometries are
represented in the BSP tree by simply discarding the degenerate yet
still producing a BspLeaf.