Fixed|BSP Builder: Space partitioner logic error resulting in mismatched sectors
When building line segments along the partitioning half-plane it is
imperative that the EdgeTip sets for the respective vertexes are
updated by inserting tips for the new line segments. Otherwise errors
will result if any of these segments are later split by subsequent
partitioning.

Closed sectors are no longer erroneously reported as "unclosed".