Fixed|BSP Builder: Space partitioner logic error resulting in duplicate line segments
When building line segments along the partitioning half-plane care
must be taken to ensure that new segments are not creating on top of
the one chosen as the partition line. Otherwise duplicate edges will
find their way into the BSP leaf geometries.

Also removed Partitioner's PartialBspLeafBuilt notification as any
malformed geometry should be considered the net result of a bug in
partitioning algorithm.