BSP Builder: Defer creation of half-edges; replaced sector selection algorithm
It is no longer necessary to produce half-edges from the outset as
the partitioning algorithm works primarily with LineSegments. Instead
construction of geometry HEdges is now deferred until such time as a
convex subspace is found (and a BspLeaf created).

The algorithm responsible for selecting the sector for a BSP leaf was
replaced with a heuristic analyzer which rates the candidate sectors
according to the number and type of referencing line segment.

Plus cleanup.