Client|Widgets: Use rich text formatting in the LogWidget
Style markup in the log entries is now taken into use in the
LogWidget. Font::RichFormat is used in FontLineWrapping and
GLTextComposer. Lines can also be indented according to the indentation
markers in the rich text.

Todo: de::Font needs to know which colors to use for the rich palette