Fixed|Resources: Misinterpretation of LINEDEF flag ML_TWOSIDED
Originally intended as a performance optimization, this flag must be
specially handled as it is not applied automatically by the original
games at map load time. Obviously there is case here for not doing
so but this feels like more of an oversight IMO.

Added Line::Side flag SDF_SUPPRESS_BACK_SECTOR which is applied when
interpreting the ML_TWOSIDED flag in id Tech 1 format maps.

The ML_TWOSIDED flag is now applied to XLine's flags and is now used
in accordance with the original games' playsim algorithms.

This fixes the holes seen in some old mods such as the first map in
Aliens TC.