BSP Builder: Installed ConvexSubspace; cleanup
Next step: Defer geometry generation until the plane has been fully
partitioned. ConvexSubspace can be used as a "staging area" for the
line segments (i.e., we don't need to build BspLeafs as soon as the
a convex subspace is found/produced).

Although we could determine the sector as soon as a convex subspace
is found it would mean duplicating logic within ConvexSubspace which
the geometry generation needs also (winding the segments).