Fixed|Line: Oversight in Line::considerOneSided()
An id Tech 1 format map with a LINEDEF sporting an unnecessary second
SIDEDEF will typically result in a degenerate BSP leaf on that side.
As such degenerates are not yet "collapsed" during the partitioning
process (and FakeRadio is still using the LineOwner rings) this case
must be specially tested for in Line::considerOneSided().