ConvexSubspace: Cleanup
The geometry generated by the space partitioner is now looking pretty
clean with no partial overlaps and only "duplicate" and/or misordered
line segments (which will be dealt with shortly).

However it is now clear that ultimately resorting to a heuristic based
method of BSP leaf sector selection is inevitable.