Client|UI: Improved integration between task bar, console, and game
The console widget is now a more integral part of the task bar. The
command line is in the task bar, and when it gains focus, the log
history opens.

When the task bar is open, Esc will now first close and unfocus the
console. The second press of Esc will close the task bar.

Shift-Esc will always move focus to the console command line.

Clicking on the LegacyWidget will clear UI focus, allowing the game
to receive events.