Fixed|LogWidget: Rendering glitch related to atlas optimization
As the entry atlas is used in a dynamic fashion, defragmentation is needed
to free up fragmented space. However, if defragmentation occurs during the
rendering of entries, the generated geometry becomes obsolete as the UV
coordinates of all entries will likely change. Therefore, LogWidget needs
to redo the visible range of entries if the atlas is optimized.