Fixed|Games: Notification about changing the current game
DD_ChangeGame is responsible for actually carrying out the changing
of the current game, so the audience must be notified only after this
function has finished running.

Todo for later: The application should be responsible for game changing.
Games should either be a simple collection that does not know the
current game, or it should be managing the current game's changing
fully (effectively merging DD_ChangeGame into it).