Refactor|Client: Lines can be composed of smaller segments
Tab stops and alignment of content between lines require that the
parts are positioned at GL drawing time rather than when the text
is wrapped and rasterized.

These changes add the concept of Segments to wrapped lines, where
each line is split into segments based on the tab stops defined with
escape sequences.

Todo: Each segment will be rasterized on a separate image on the atlas
so that they can move freely during GL composition.