Client|Widgets: Blurred GuiWidget background, fine-tuning
A GuiWidget background can now be configured to be blurred. The widgets
behind the blurred widget are drawn onto a texture and blurred using
a shader.

Note about performance: each GuiWidget does the effect independently,
so it is not a good idea to use it everywhere. Particularly when
LegacyWidget is not intended to be drawn more than once per frame.

The ring zero UI background is now drawn using a LabelWidget instead
of the old UI code.

Todo: LegacyWidget is currently disabled as it messes the OpenGL state
in some way, causing other widgets to go missing.