API: Revised API for map changes
Game plugins now perform a map change by calling the API function
P_MapChange(). Once the map change has been completed the engine
will call the FinalizeMapChange function exported by the game (if
any) to allow it perform any post change finalization it needs at
this time.

It is no longer necessary for the game plugin to inform the engine
when a map change is about to occur (which was previously done by
calling R_SetupMap() with the DDSMM_INITIALIZE mode argument).

Similarly it is no longer necessary for the game plugin to inform
the engine once it has completed post-change finalization (which
was previously done by calling R_SetupMap() with DDSMM_FINALIZE
mode argument).

R_SetupMap() should now only be called after loading a save state
to inform the engine to perform setup for the map again.