World: Reordered events in the mapChange() procedure
The aim being to group and order the initialization of subsystems and
their data according to their concerns and also to maximize what is
possible during the map change finalization on game side.

Naturally only so much can be achieved like this as there are a lot
of inter dependencies which prevent an ideal ordering.

Hopefully I didn't break anything...