Refactor: Moved the current game and game change audience to App level
Moving these from the Games collection up to app level is important
as this change as this significantly alters game change semantics.

However, presently de::App can not implement these as it would mean
relocating the Games collection to libdeng2 (which is not feasible
yet as Game depends on the FS1 file system). So for now these are
implemented as globals.