Map: Cleaned up interface to the binary space partitioner
As the BSP data elements are now produced using data and mechanisms
provided by the map it is no longer logical to "black box" the whole
build process by encapsulating behind a high-level BspBuilder class.

The BspBuilder wrapper was dumped and Reporter relocated to World.

Also reworked the interface to support building a BSP using only a
subset of the map's lines in preparation for future work which will
begin post 1.11 for Doomsday 1.12

Todo: Reconnect the Reporter to the map conversion process so that
any issues or problems encountered are logged.