Client|UI: Default binding for toggling taskbar+console
This commit replicates the old behavior of the special console
activation key using bindings: a Tilde without a Shift modifier toggles
the taskbar+console.

This also shows the first steps of migrating the bindings system to
work together with widgets instead of relying on context (de)activation.
The plan is that widgets will check for bindings associated with them
as part of the normal widget event dispatch.

The "console" context bindings now override ALL other event handling
in the ConsoleCommandWidget.