Fixed|Client|LineEditWidget: Assert failed when navigating history rapidly
The editor's history navigation requires line-wrapped content. However, the
wrapping was not being updated until the next frame was drawn. Thus, if one
tried to navigate two steps during one frame, the wraps were not up to date
when they were needed.

Now the wraps are updated immediately if the editor widget is part of a
widget tree.