Fixed|Client: Crash and mishandling of failed entry allocations
When the entry atlas is so full that an allocation of an entry fragment
fails, LogWidget did not handle the situation correctly. Now if this
occurs, all out-of-view items are released to make room on the atlas
and one extra attempt is made to allocate the visible entries.

Also increased the maximum size of the entry atlas (ideally this would
be display mode dependent).

A crash was occurring because all entries were zapped from the atlas
during the update of an entry, confusing GLTextComposer.