Client|ClientWindow: FPS counter as a notification
This commit brings back the FPS counter as a notification widget owned
by ClientWindow. When "Show FPS" is selected in the DE menu, the
window sets Config.window.main.showFps to True and the FPS counter
widget is shown in the main window's notification area.

The counter notification is dismissed by selecting "Hide FPS" in the
DE menu.