Optimize|Client|UI: More intelligent text line alloc/release on atlas
The line edit widget was releasing and allocating its entire contents
even though a single character was added to one of its lines. This
incurred a bit performance hit when there was a large number of lines
in the editor.

Now each text segment is more carefully tested in GLTextComposer to
see if it is really necessary to release and reallocate it.