Homepage: Revised front page design; use jQuery XML parsing for build events
Sadly it seems that phpBB's Atom feed is completely uncached, so for
the news feeds we'll use Google's Feed API service (which implements
the cache that phpBB should have...).

However this means that we can't extract the feed "source" URI from
the JSON response which Google serves (in the case of an Atom feed
for a subforum phpBB encodes the source using the 'domain' attribute
of the category tag, which is not interpreted by Google's service).