Committed the Deng phpBB forum skin to the code repository
This is the phpBB 3.0.8 forum skin as seen on dengine.net

Note that as this skin inherits from the default prosilver (version
3.0.8) -- not all template and image files are included as several
are unchanged.

What is included are all CSS style rules pertinent (does not depend
on prosilver), all custom images and page templates.

It should also be noted that for the most part the forum styling is
inherited from the "global" dengine.net style sheet.

Todo for later: Continue clean up and restructuring of the rules and
all page templates. Organization here leaves a lot to be desired.