Fixed|Resources|BSP Builder: Uncaught exception loading HYMN_BV2.wad E1M6
This map contains several instances of a particularly nasty construct
from the point of view of a BSP builder -- two-sided lines facing out
into the void. Such constructs will invariably result in degenerate
BSP leafs.

Should such a degenerate (i.e., zero-area) leaf consist of multiple
line segments the angle-sorted order cannot be trusted and so overlap
splitting should not be attempted.

For now we will simply skip splitting of each segment if the would-be
coordinates are equal to those of either of the segment's vertexes.

Todo for later: Implement an algorithm to detect and exclude these at
an earlier point.