Fixed|BSP Builder: Ensure BSP output is deterministically predictable
The set of map lines provided to the space partitioner is in the form
of a QSet to ensure that no duplicate line references are passed in.
However, the BSP builder should ensure that given the same set of lines
the output is deterministically predictable. Therefore we must sort the
input line set by unique in-map index prior to partitioning.