Fixed|libdeng2|Time: Normal time takes precedence over high-performance time
By default, Time instances are initialized with both a QDateTime and
a high-performance time. This means both are serialized persistently,
however the high-performance time is relative to the start time of
the previous session and therefore cannot be directly compared
against in a subsequent session.

It is still valuable to keep the relative high-performance time in
serialization, e.g., for serializing Animation instances.