MapElement: World map attribution and parent/child ownership hierarchy
MapElement now implements a mechanism for modeling a parent/child
ownership hierarchy and built-in property value delegation should a
parent element be configured.

This mechanism is now employed by many of the core map element types,
including the Surface->Plane->Sector and Surface->Line::Side->Line
hierarchies.

Each MapElement may also now be attributed to a world Map instance.

The new functionality was made use of to clean up cases where a map
element needs to refer to another element of the same map (which, is
not linked in some other way).