Fixed|libcommon|64-bit: Crash when binding controls in menu
Fixed passing of the symbolic event string pointer in the event_t
structure. Reassembly of the 64-bit pointer failed if ev->data1
was negative in its 32-bit integer range.

Todo for later: It isn't a great idea to pass 64-bit pointers like
this, why not just add a pointer member to event_t?