libcommon: Place message log below map title
If the map title is visible, either due to the start of a map or
because the automap is open, the HUD message log is moved downwards
a little so it doesn't overlap.

Thanks to Vermil for reporting the issue.

Also fixes the issue where the message log would not be drawn at all
when the title is shown in the beginning of a map.