Server|Fixed: Transmission of malformed DT_SIDE deltas
As the sections of a Line::Side are now optional the lineFlags field
of a dt_side_t was not handled properly when comparing a delta for
the back side of one-sided line. In this situation the register was
initialized correctly but Sv_RegisterCompareSide then compared vs
a zeroed flags field. Consequently the server transmitted a malformed
DT_SIDE delta.

This commit fixes the "material y-axis pegging" issues in multiplayer.