BiasSource|Optimize: Do not observe sector lighting changes in the void
While origin of the source lies in the void tracking of sector light
level changes is now disabled as this resulted in unnecessary change
notifications filtering through to surfaces.

Todo: It appears there are some latent light contributions at vertex
level which aren't be interpolated away when the source is occluded
on either the X|Y plane or the Z-axis.