Fixed|Shadow Bias|BiasSurface: Latent bias lighting contributions when surface loses all sources
When recalculating per-vertex bias lighting the special case of a
surface suddenly loosing all light contributions was not handled
correctly. This resulted in latent contributions from sources which
no longer affect the surface continuing to do so.

Todo: BiasSurface::Instance::evalLighting() should be revised as it
is now obvious that a previous refactoring effort to minimize the
effect of source lighting changes was not completed.