Refactor|BiasSurface: Integrated 'latent' light contribution concept at BiasSurface level
As light contributions are no longer referenced by the index of the
BiasSource, they no longer erroneously propagate to another source
when another is deleted.

Todo: The vertexes are still working under the assumption that they
must manage this concept, with the net result being a conflict come
interpolation time.