BiasSurface|Optimize: Vertexes are no longer concerned where light contributions originate
As 'latent' contributions are now managed at surface level there is
no longer a need for vertexes to be aware of the source of each of
it's light contributions. Also, the 'unseen' state is now dealt with
by the surface.

This reduces the storage requirement for each vertex by 27 bytes.