libgui|libdeng2|Client: Added a way to reset indent in rich format
In the rich format, the indent reset is inserted as an empty range,
so that it can be applied as the last thing on a line. If it were
included in the range that includes a newline, FontLineWrapping would
skip it as it only operates on whatever is between newlines.

The escape sequence for reseting indent is '<'.

EscapeParser was adjusted accordingly (< was previously thought to
begin a long sequence terminated by >).