BiasSurface: Further revised the internal design of this component
In this iteration vertex illumination is fully decoupled from the
"surface", whose main job now is tracking changes and routing through
update notifications from sources to vertexes.

The next step is separating the illumination form the surface so it
can again be managed with an external buffer (the difference being
that now these are thinking VertexIllum instances).