Fixed|Client|Resources: Hires texture suppression in release build
It appears that when declaring flags with Qt it is vital to ensure
that one takes the additional step of declaring operators for said
flags. No warnings are issued by the compiler should one fail to do
so and the result is miscompiled code.

I'm no API design guru but I personally find having to use both of
Q_DECLARE_FLAGS and Q_DECLARE_OPERATORS_FOR_FLAGS a real PITA.