World|Map|DMU API: Removed ability to traverse line segments of a BSP leaf using DMU
Line segments should no longer be thought of as an element of the
map geometry. In the future the half-edges of primary map geometry
will be made available via DMU. Segments are now on their way out.

When the time comes to implement a textured automap the engine will
provide suitable (clean) prebuilt geometry for a visited BSP leaf.