Refactor|Client|UI: Added ProceduralImage base class; use it in LabelWidget
Introducing the concept of procedural images brings a lot of
flexibility to labels and other widgets derived from it (e.g., button).

A procedural image can generate any geometry instead of just a
simple 2D quad with a texture.

LabelWidget now uses ProceduralImage internally to draw the Image of
the label.