World|Line: Line::Side has ownership Segments; basic API for management
In the future this mechanism will be replaced entirely, so for now
this simple implementation will suffice.

Todo for later: Both BspLeaf and Line::Side now implement a similar
mechanism for management of "subelement geometries". This mechanism
should be generalized with a uniform (likely URI based) method of
addressing geometries (libdeng2's Bank seems to be a good fit).