Refactor|libcommon: Command "setmap" is now an alias of "warp"
The former is no longer necessary as management of game sessions and
game state is now more logical (in particular the saved ACS state in
Hexen is now managed as part of the map change process once the game
session begins).

When executed on server-side a warp command now respects the current
value of "server-game-skill" when (implicitly) starting a new game
session.

This needs testing (all games, multiplayer map cycle, etc...).