Fixed|World: Fatal error attempting to play icarus.wad MAP16
This map contains certain mapping constructs which we now interpret
into 'extra' mesh geometries. Recently the way these geometries are
constructed was changed so that segments are generated only for the
edges of map lines. It appears that I mistakenly overlooked the need
to update the line-of-sight algorithm accordingly (the DMU API also
required an update).