Fixed|Map Renderer: Fatal error firing the BFG in DOOM mod "Astrostein"
This mod uses truncated Patch format images as a means to suppress
drawing of the sprite (ugh...). While the texture resource loader
and material system handle this situation the map renderer failed
to abort when "projecting" the sprite into the world.

Todo for later: Presently this will result in repeated attempts to
load the invalid texture. We need a mechanism which flags invalid
resources when a load attempt fails.