World|Sector: Permanently link sector clusters bound by self-referencing lines
Theoretically at least, this relatively simple algorithm should fix
the majority of map hacks utilizing self-referencing lines (including
specializations such as "deep water"). The key to solving this is the
the revised BSP geometry construction and the implicit preconditions
which the clustering inherently ensures.

Todo: This needs a lot of testing and the implementation cleaning up.
Todo: Document and provide proof of how and why this works.