UI|Client: Various widget improvements; cleanup
ProgressWidget's default mode is to align content normally rather
than by image only, as the latter is only really needed in
BusyWidget.

LabelWidget determines the maximum available text width better:
the constraints set on the label image must be taken into account.

PopupWidget constrains the placement of the popup so it doesn't
go outside the view with the Down opening direction.