Refactor|libgui|Client: Untabbed lines among tabbed lines
It is now possible to set the tab stop to -1 to indicate that the
line should not be tabbed at all. The objective is to allow reseting
the tab stops within a longer piece of content, continuing with a
new set of stops.

Refactored FontLineWrapping and GLTextComposer to allow processing
tabbed lines in distinct ranges.