Updater|Config|Client: Updater's default config moved to appconfig.de
Previously this was handled by native code in UpdaterSettings, but
now Doomsday Script has all the information available to make the
proper decisions about the Updater's default settings.

Having this in appconfig.de ensures that the normal de::Config
(re)configuration mechanism can be applied to the updater as well.