Fixed|World|Sector: Re-evaluate sector cluster linking following containment testing
In this situation there may still be another map hack which contains
the cluster (which, contains the nested map hack). To resolve this
we'll repeat the visual plane remapping process.

Todo for later: Re-implement this algorithm to avoid recursion. This
should wait until after "missing texture" map hack support, however.