Map Renderer|Client: Use sector clusters for plane ambient light properties
When drawing the geometry of a map plane ambient light properties
should be determined by the sector cluster (as these may be mapped
from another sector).

For now this is implemented with a temporary map renderer draw state
change as it is not yet clear whether this should be a fundamental
part of the map lighting model.