Map Renderer|Client: Revised mechanism for generating Lumobjs from sprites
This logic is now split into two distinct phases:

Phase 1 involves the generation of Lumobjs from a sprite and at this
point are not linked in the map. Properties of which are configured
accordingly in "local space" relative to the sprite from which it is
generated.

Phase 2 takes the generated Lumobjs as input, further (re)configures
them according to the mobj they will be used with (e.g., translating
their origin and applying property overrides from Light definitions)
before then inserting them into the Map.