Map Renderer|SurfaceDecorator|Client: SurfaceDecorator is responsibility for applying material changes
When the properties of a material are changed directly or animated,
SurfaceDecorator assumes responsibility for scheduling and applying
such changes to the map surfaces in it's charge.

Note that when surface changes (say, a new material is assigned) it
is the responsibility of the user to notify the relevant decorator.