UI|Task Bar: Various tweaks to focus/mouse trap behavior
When the sidebar is open, the task bar will behave a bit differently.
While the "taskbar" command will still always open the task bar and
set focus to the command line, Shift-Esc will now just open the task
bar, as the assumption is that the user is interested in accessing
the sidebar editor rather than the command line.

Clicking on nothing or a panel background will now clear the input
focus.