Refactor|World|Plane|Surface: Disambiguate "visual height" wrt planes
Since the introduction of mapped sector planes the "visual height"
of a plane could be interpreted with two very different meanings
depending on the context.

Sector cluster allows mapping of planes for visual effect only and
so the term "visual" is now overloaded (previously this identified
the interpolated "smooth" height of a plane).

To disambiguate, Plane's API now uses the term "smoothed" and for
uniformity, Surface applies the same terminology when identifying
the smoothed material origin.