Refactor|Map Renderer|Client: Removed redundant Surface::DecorSource
Now that the need for surface decoration updates is tracked and the
updates done in a timely fashion -- the old DecorSource abstraction
was unnecessary.

When SurfaceDecorator redecorates a Surface the old decorations are
cleared before then generating new ones according to the material
definition, which Surface then takes ownership of.

Todo: Cleanup