Fixed|Hexen: Excessive light for Hexen's Mage Frost Shard FX
The old DDMF_LIGHTSCALESHIFT API flag was removed. Previously this
was used to dynamically scale a luminous object on a frame by frame
basis. Lights are being transitioned to a model where their lifetime
spans multiple render frames, so dynamically manipulating such values
via a public API is not really compatible.

The only time this is used currently is with the Mage's Frost Shards
in Hexen and here this value does not change between frames.

A definition can be used to replace this if necessary.