Fixed|libgui|Font: Use normal weight ascent for other weights
Some fonts have a different ascent for the different weights. If a
widget used a font with such a weight, it would not align with
normal-weight text in other widgets.

Now all weights of a font use the normal ascent, the assumption being
that the ascents of the weights do not differ too greatly (as the
font point size is the same).