Fixed|BSP Builder: Space partitioner incorrectly assigned sectors adjoining map hacks
When building line segments along the partitioning half-plane it is
not simply the self-referencing state of the intercepted map line
which is important when choosing a sector. Rather, it is the lines
from which a sector will be chosen whose self-referencing state is
important.

The older "previous sector" kludge which achieved limited success at
fixing this has now been removed.

This commit fixes a variety of maps such as; not fully deep water in
TNT MAP31; sunken barrels in AV MAP15; dark nukege in KSutra MAP13.