Fixed|World|Sector|Client: Crashes in various Hexen maps when mapping sector planes
The Hexen IWAD is littered with maps containing severely broken map
geometry, such as one-sided line constructs facing outward into the
void. If such a construct partly forms a convex subspace this will
produce a non-degenerate BSP leaf and which may be "attached" to the
back of otherwise valid geometry.

Unfortunately we can't simply remove such geometries from the map at
load time because the playsim may depend on their presence to ensure
that line specials (for example) work correctly. Consequently these
must be tolerated and carefully working around them.

If such a construct is encountered when classifying sector clusters
for visual plane mapping, the cluster should be marked such that all
mapping is canceled. The assumption being that if the map author did
intend to use a map hack here then they would fixed the geometry in
any case.