Fixed|libcommon: Behavior of sector type 10 "close in 30 seconds"
When a mobj blocks the path of the plane preventing from completing
its mission of sealing off an area -- vanilla behavior is to invert
the move and instead raise continuously until it wraps around.

We'll ape ZDoom in this situation by bouncing the plane back up to
it's original height and then stopping.