Fixed|TextDrawable|DocumentWidget: Updating text (e.g., autocompletion list)
DocumentWidget was jumping the gun and updating the text before
TextDrawable had finished the background wrapping task of the new
content. This was a regression that occurred when TextDrawable was
switched to a double-buffered model.

Now TextDrawable::update() will not return true until the background
wrapping is complete, after which DocumentWidget will update its
content size and visible lines.