Cleanup|World|BspLeaf: Removed all obsolete BspLeaf accessor methods
A BSP leaf should not be thought of as a "subsector" as this produces
too strong of a binding with the much higher level "sector" concept.
(Also resulting in a wide and heavy API at BspLeaf level.)

If one wants sector and/or plane properties at a BSP leaf, the sector
cluster is now dereferenced explicitly. (Which should help to guide
future map data development in the right direction.)