World|Map: Allow linking mobjs to a sector of a degenerate BSP leaf
There appears to be some confusion as to what it means for a mobj to
be "sector linked". Specifically, whether a mobj in a sector list is
"within" the bounds of that sector according to line geometry of the
map, or, if only in a BSP leaf attributed to that sector.

Essentially DOOM confuses the definition of a subsector, exchanging
"BSP half-space" with "convex subspace" depending on the context.
The name "subsector" probably didn't help to resolve it either.

For compatibility purposes, when linking mobjs to sectors we should
apply the BSP half-space definition for now.

Todo for later: Most likely there are more mixups and/or incorrect
assumptions in any code which uses the sector mobj lists...