Fixed|BSP Builder|ConvexSubspace: Failed to partition a map with overlapped single sided lines
When two or more single sided lines are overlapping and reference the
same sector, depending on how the map is partitioned it may result in
a degenerate set of line segments. In this situation, separating the
discordant segs into an "extra" mesh geometry may mean we are left
with fewer than three to produce the primary convex polygon geometry
for a BSP leaf. In which case, we shouldn't attempt to assign a non-
convex polygon.

BLUDWRKS.wad once again loads (and very nearly rendered correctly).