Fixed|Resources: Fatal error attempting to load a map containing lines with no sides
Now that the engine is using the half-edge data structure and sector
clusters for the majority of tasks, such lines can be supported.

I've tested a variety of such maps and everything appears to work as
expected. As usual however, if you encounter an issue in such a map
the please do report the problem.

See: [bugs:#965]